Feat: Chapter 7: Using Constraints, Filtering, Searching, Ordering, and Pagination
We understood the importance of paginating results, and we configured and tested a global limit/offset pagination scheme with the Django REST framework. Then, we created our own customized pagination class to make sure that requests weren't able to acquire a huge amount of elements in a single page

We configured filter backend classes, and we added code to the models to add filtering,
searching, and ordering capabilities to the class-based views. We created a customized filter, and we made requests to filter, search, and order results, and we understood how everything worked under the hood. Finally, we used the browsable API to test pagination, filtering, and ordering

In the models.py added unique constraint to each model and new migration with these constraints applied

In the settings.py the REST_FRAMEWORK dictionary defined with DEFAULT_PAGINATION_CLASS and PAGE_SIZE for pagination feature (offset/limit) and DEFAULT_FILTER_BACKENDS tuple for filter enumeration in generics.ListCreateAPIView: classes

In the new custompagination.py file defined upper bound for limit in each request to prevent a huge amount of data of default undefined upper bound

In the requirements.txt the 'django-filter' lib added and registered-applied in settings.py

In the views.py in each class which is ListCreateAPIView subclass defined filterset_fields, search_fields and ordering_fields. For CompetitionList class-based view defined custom CompetitionFilter