Feat: Chapter 10: Automating Tests
In this chapter, we learned to write unit tests for our RESTful Web Service. We installed the necessary packages and made the appropriate configurations to work with the modern and popular pytest unit test framework. Then, we wrote our first round of unit tests for the
RESTful Web Service related to different scenarios with drone categories.

We worked with the different options for the pytest command to discover and run unit tests in the default mode, the increase verbosity mode, and the disable capture mode. We
understood how to combine pytest with the testing classed provided by the Django REST
framework.

Finally, we wrote additional unit tests for the RESTful Web Service related to different scenarios with pilots and the token authentication requirements for specific requests. We are able to continue adding tests for our RESTful Web Service with all the things we have learned.

The pytest library installed and configured in pytest.ini. Also, requirements.txt has been updated

In the drones/tests.py created DroneCategoryTests and PilotTests classes which inherit APITestCase